refactor pmmr, split out readonly_pmmr and backend
