+++ core/src/core/mod.rs
@@ -23,7 +23,6 @@ pub mod hash;
 pub mod id;
 pub mod merkle_proof;
 pub mod pmmr;
-pub mod readonly_pmmr;
 pub mod transaction;
 pub mod verifier_cache;
