+++ core/src/core/pmmr/pmmr.rs
@@ -12,30 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//! Persistent and prunable Merkle Mountain Range implementation. For a high
-//! level description of MMRs, see:
-//!
-//! https://github.
-//! com/opentimestamps/opentimestamps-server/blob/master/doc/merkle-mountain-range.
-//! md
-//!
-//! This implementation is built in two major parts:
-//!
-//! 1. A set of low-level functions that allow navigation within an arbitrary
-//! sized binary tree traversed in postorder. To realize why this us useful,
-//! we start with the standard height sequence in a MMR: 0010012001... This is
-//! in fact identical to the postorder traversal (left-right-top) of a binary
-//! tree. In addition postorder traversal is independent of the height of the
-//! tree. This allows us, with a few primitive, to get the height of any node
-//! in the MMR from its position in the sequence, as well as calculate the
-//! position of siblings, parents, etc. As all those functions only rely on
-//! binary operations, they're extremely fast.
-//! 2. The implementation of a prunable MMR tree using the above. Each leaf
-//! is required to be Writeable (which implements Hashed). Tree roots can be
-//! trivially and efficiently calculated without materializing the full tree.
-//! The underlying Hashes are stored in a Backend implementation that can
-//! either be a simple Vec or a database.
-
 use std::marker;
 use std::u64;
 
@@ -43,69 +19,14 @@ use croaring::Bitmap;
 
 use core::hash::Hash;
 use core::merkle_proof::MerkleProof;
-use core::readonly_pmmr::ReadonlyPMMR;
+use core::pmmr::{Backend,ReadonlyPMMR};
 use core::BlockHeader;
 use ser::{PMMRIndexHashable, PMMRable};
 use util::LOGGER;
 
+
 /// 64 bits all ones: 0b11111111...1
 const ALL_ONES: u64 = u64::MAX;
-
-/// Storage backend for the MMR, just needs to be indexed by order of insertion.
-/// The PMMR itself does not need the Backend to be accurate on the existence
-/// of an element (i.e. remove could be a no-op) but layers above can
-/// depend on an accurate Backend to check existence.
-pub trait Backend<T>
-where
-	T: PMMRable,
-{
-	/// Append the provided Hashes to the backend storage, and optionally an
-	/// associated data element to flatfile storage (for leaf nodes only). The
-	/// position of the first element of the Vec in the MMR is provided to
-	/// help the implementation.
-	fn append(&mut self, position: u64, data: Vec<(Hash, Option<T>)>) -> Result<(), String>;
-
-	/// Rewind the backend state to a previous position, as if all append
-	/// operations after that had been canceled. Expects a position in the PMMR
-	/// to rewind to as well as bitmaps representing the positions added and
-	/// removed since the rewind position. These are what we will "undo"
-	/// during the rewind.
-	fn rewind(&mut self, position: u64, rewind_rm_pos: &Bitmap) -> Result<(), String>;
-
-	/// Get a Hash by insertion position.
-	fn get_hash(&self, position: u64) -> Option<Hash>;
-
-	/// Get underlying data by insertion position.
-	fn get_data(&self, position: u64) -> Option<T>;
-
-	/// Get a Hash  by original insertion position
-	/// (ignoring the remove log).
-	fn get_from_file(&self, position: u64) -> Option<Hash>;
-
-	/// Get a Data Element by original insertion position
-	/// (ignoring the remove log).
-	fn get_data_from_file(&self, position: u64) -> Option<T>;
-
-	/// Remove Hash by insertion position. An index is also provided so the
-	/// underlying backend can implement some rollback of positions up to a
-	/// given index (practically the index is the height of a block that
-	/// triggered removal).
-	fn remove(&mut self, position: u64) -> Result<(), String>;
-
-	/// Returns the data file path.. this is a bit of a hack now that doesn't
-	/// sit well with the design, but TxKernels have to be summed and the
-	/// fastest way to to be able to allow direct access to the file
-	fn get_data_file_path(&self) -> String;
-
-	/// Also a bit of a hack...
-	/// Saves a snapshot of the rewound utxo file with the block hash as
-	/// filename suffix. We need this when sending a txhashset zip file to a
-	/// node for fast sync.
-	fn snapshot(&self, header: &BlockHeader) -> Result<(), String>;
-
-	/// For debugging purposes so we can see how compaction is doing.
-	fn dump_stats(&self);
-}
 
 /// Prunable Merkle Mountain Range implementation. All positions within the tree
 /// start at 1 as they're postorder tree traversal positions rather than array
